requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
customtkinter==5.0.3
darkdetect==0.8.0
filelock==3.12.0
future==0.18.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
keyboard==0.13.5
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8
numpy==1.23.5
opencv-python==4.6.0.66
outcome==1.2.0
pefile==2022.5.30
Pillow==9.3.0
pycparser==2.21
pygame==2.1.3.dev8
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySocks==1.7.1
pyttsx3==2.90
pytube==15.0.0
pywin32==305
pywin32-ctypes==0.2.0
regex==2022.10.31
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
torch==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.6.2
urllib3==1.26.13
wsproto==1.2.0